refactor: replace relative paths for headers with full names.
Currently no header can be moved without worrying about which source files include it.
Having ../.. everywhere is not self-describing. Using full names makes it very clear
what directory the header file belongs to.

This change involves adding the -I flag to CFLAGS in the Makefile to find the files
in the cread directory.